Allow custom mapping directory in backward.py
This is a rough implementation that allows user to choose where to look
for the mapping files. The user chooses the directory using the -mapdir
option. The user directory **replaces** the default mapping directory,
which might not be ideal.

Being able to choose the mapping directory allows to use custom mappings
without tempering with the MartiniTools directory. It also allows to
write mappings on the fly using a script without the risk of loosing
defaults mappings.